requirements.txt
streamlit-authenticator
pyairtable
streamlit-echarts
pandas
streamlit-drawable-canvas
altair
streamlit-extras